refactor(PolicyBase): rename PolicyBase to DefaultPolicy
