+++ lib/amqp_client.js
@@ -23,7 +23,7 @@ var EventEmitter = require('events').EventEmitter,
 
     Translator = require('./adapters/translate_encoder'),
 
-    PolicyBase = require('./policies/policy_base'),
+    DefaultPolicy = require('./policies/default_policy'),
     EHPolicy = require('./policies/event_hub_policy'),
     SBQueuePolicy = require('./policies/service_bus_queue_policy'),
     SBTopicPolicy = require('./policies/service_bus_topic_policy'),
@@ -40,7 +40,7 @@ var EventEmitter = require('events').EventEmitter,
  * need to speak AMQP-specific (e.g. to set a filter to a described-type), you can use node-amqp-encoder and the
  * AMQPClient.adapters.Translator adapter to convert it to our internal types.  See simple_eventhub_test.js for an example.
  *
- * Configuring AMQPClient is done through a Policy class.  By default, PolicyBase will be used - it assumes AMQP defaults wherever
+ * Configuring AMQPClient is done through a Policy class.  By default, DefaultPolicy will be used - it assumes AMQP defaults wherever
  * possible, and for values with no spec-defined defaults it tries to assume something reasonable (e.g. timeout, max message size).
  *
  * To define a new policy, you can merge your values into an existing one by calling AMQPClient.policies.merge(yourPolicy, existingPolicy).
@@ -68,12 +68,12 @@ var EventEmitter = require('events').EventEmitter,
  * If, on construction, you provide a uri and a callback, I will immediately attempt to connect, allowing you to go directly from
  * instantiation to sending messages.
  *
- * @param {PolicyBase} [policy]     Policy to use for connection, sessions, links, etc.  Defaults to PolicyBase.
+ * @param {DefaultPolicy} [policy]     Policy to use for connection, sessions, links, etc.  Defaults to DefaultPolicy.
  * @constructor
  */
 function AMQPClient(policy) {
   // Make a protective copy
-  this._originalPolicy = u.deepMerge(policy || PolicyBase);
+  this._originalPolicy = u.deepMerge(policy || DefaultPolicy);
   this.policy = u.deepMerge(this._originalPolicy);
   this._connection = null;
   this._session = null;
@@ -120,15 +120,15 @@ AMQPClient.adapters = {
 
 
 /**
- * Map of various pre-defined policies (including PolicyBase), as well as a merge function allowing you
+ * Map of various pre-defined policies (including DefaultPolicy), as well as a merge function allowing you
  * to create your own.
  */
 AMQPClient.policies = {
-  'PolicyBase': PolicyBase,
+  'DefaultPolicy': DefaultPolicy,
   'EventHubPolicy': EHPolicy,
   'ServiceBusQueuePolicy': SBQueuePolicy,
   'ServiceBusTopicPolicy': SBTopicPolicy,
-  merge: function(newPolicy, base) { return u.deepMerge(newPolicy, base || PolicyBase); },
+  merge: function(newPolicy, base) { return u.deepMerge(newPolicy, base || DefaultPolicy); },
   utils: putils
 };
