+++ lib/connection.js
@@ -144,7 +144,7 @@ var EventEmitter = require('events').EventEmitter,
  *
  * Also could be DISCARDING if an error condition triggered the CLOSE
  *
- * @param connectPolicy ConnectPolicy from a PolicyBase implementation
+ * @param connectPolicy ConnectPolicy from a DefaultPolicy implementation
  * @constructor
  */
 function Connection(connectPolicy) {
