+++ lib/policies/service_bus_policy.js
@@ -2,7 +2,7 @@
 
 var constants = require('../constants'),
     u = require('../utilities'),
-    PolicyBase = require('./policy_base');
+    DefaultPolicy = require('./default_policy');
 
 module.exports = u.deepMerge({
   senderLink: {
@@ -45,5 +45,5 @@ module.exports = u.deepMerge({
       }
     }
   }
-}, PolicyBase);
+}, DefaultPolicy);
